Move the start of the service to start of config change
